Add a ? to end of URL
The reason for this is that discord currently makes a new request to the
server if and only if the URL ends in a ?
Without a ? and also with anything coming after the ? discord will cache
the result which makes the site harder to use for shitposting. Without
caching however we can send whatever we want back to discords thumbnail
generator.
So every time a user links the site the link should be followed by a
single ?